Restructure things a little bit
